Improved helper script

+++ prepare-ghz.sh
@@ -1,4 +1,16 @@
 #!/bin/bash
+
+FILE="95131"
+MAPS_FOLDER="$HOME/projects/maps"
+MAP_START_POS="50.2836424,11.6190666"
+MAP_START_ZOOM="17"
+BBOX="50.26696,11.5864563,50.3011828,11.6777802"
+
+INPUT_FILE="$MAPS_FOLDER/${FILE}.osm"
+GH_FOLDER="$MAPS_FOLDER/${FILE}-gh/"
+MAP_FILE="$GH_FOLDER/${FILE}.map"
+
+
 if [ "$1" == "-h" ]; then
   echo "Pre-Requirements:"
   echo " * osmosis with map writer plugin"
@@ -6,12 +18,14 @@ if [ "$1" == "-h" ]; then
   echo " 1. Get an .osm-file"
   echo " 2. Edit the variables in this script"
   echo " 3. Execute"
-  echo " 4. "
   echo ""
   echo "Related resources:"
   echo " * https://github.com/graphhopper/graphhopper/blob/master/docs/android/index.md#maps"
   echo " * https://github.com/mapsforge/mapsforge/blob/master/docs/Getting-Started-Map-Writer.md"
-  echo " * "
+  echo ""
+  echo "Options:"
+  echo " -i  : Installs osmosis and the required plugin"
+  echo " -p  : adb pushes the created files"
   exit 0
 fi
 
@@ -23,15 +37,12 @@ if [ "$1" == "-i" ]; then
   exit 0
 fi
 
-FILE="95131"
-MAPS_FOLDER="$HOME/projects/maps"
-MAP_START_POS="50.2836424,11.6190666"
-MAP_START_ZOOM="17"
-BBOX="50.26696,11.5864563,50.3011828,11.6777802"
-
-INPUT_FILE="$MAPS_FOLDER/${FILE}.osm"
-GH_FOLDER="$MAPS_FOLDER/${FILE}-gh/"
-MAP_FILE="${INPUT_FILE::-4}.map"
+if [ "$1" == "-p" ]; then
+  for i in ${GH_FOLDER}*; do 
+    adb push $i /sdcard/Download/graphhopper/maps/${FILE}-gh/$(basename $i)
+  done
+  exit 0
+fi
 
 rm -rv $GH_FOLDER
 ./graphhopper.sh -a import -i $INPUT_FILE
